deprecated `usbd_control.c` and merge its functionality into `usbd.c`
